third_party/WebKit/LayoutTests/external/wpt/html/browsers/browsing-the-web/scroll-to-fragid/fragment-and-encoding-expected.txt

This is a testharness.js-based test.
PASS U+00FF should find U+00FF
PASS Percent-encoded UTF-8 BOM should find U+FEFF as BOM is not stripped when decoding
FAIL %FF should not find U+00FF as decoding it gives U+FFFD assert_equals: #%FF expected 0 but got 9416
Harness: the test ran to completion.
